InferenceServer/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
colourmap==0.1.1
constantly==15.1.0
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.0
incremental==21.3.0
joblib==1.1.0
json5==0.9.6
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pca==1.5.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
Twisted==21.7.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wget==3.2
wrapt==1.13.3
zipp==3.6.0
zope.interface==5.4.0

